requirements.txt
streamlit
gspread==6.1.2
oauth2client==4.1.3
pytz
openai==0.28.0               # OpenAI API client
langchain==0.1.0             # LangChain for LLM chains
python-dotenv
Flask
Flask-Cors
langchain-groq
